src/requirements.dev.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.0
Django==4.2.4
django-allauth==0.56.1
django-cleanup==8.0.0
django-phonenumber-field[phonenumberslite]==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
filelock==3.12.3
identify==2.5.27
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
phonenumberslite==8.13.20
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.5
